refactor: extract shared route/filter registrar
Move route + filter registration logic into HeliosRouteRegistrar so
production (HeliosApp.setup) and test harness (makeTestApp) share the
same code path. Eliminates duplication flagged in PR #9 review.

All 13 tests pass.
